+++ directions/employee.md
@@ -1,3 +1,9 @@
 # Managing Employees
 
 - On the Employee's page, the user can view a list of all of the employees along with their department name.
+<!-- Video of user licking employee navebar from home -->
+- From this page a user can add an employee using the "Add Employee" link as well as 
+- When a user clicks on an employee's name they are taken to a page that shows the details for that employee
+
+<!-- click employee name -->
+
