+++ hrapp/views/employees/data_manager/__init__.py
@@ -1,0 +1,4 @@
+from .computers import get_employee_computers, get_availible_computers
+from .programs import get_employee_training, get_all_trainings, assign_program
+from .employees import add_employee, edit_employee, get_employee, get_all_employees
+from .departments import get_all_departments
