--- hrapp/views/employees/data_manager/__init___.py
@@ -1,1 +1,0 @@
-from .get_all_employees import all_employees
