+++ hrapp/views/employees/data_manager/computers/__init__.py
@@ -1,0 +1,2 @@
+from .employee_computer import get_employee_computers
+from .get_availible_computers import get_availible_computers
