+++ hrapp/views/employees/data_manager/computers/employee_computer.py
@@ -1,9 +1,10 @@
 import sqlite3
-from ...connection import Connection
+from ....connection import Connection
 from hrapp.models import model_factory, EmployeeComputer
 def get_employee_computers(employee_id):
     """
-    This function gets the active computer a user has from hrapp_employeecomputer
+    This function takes the passed in employee_id to 
+    get the active computer a user has from hrapp_employeecomputer
     """
     with sqlite3.connect(Connection.db_path) as conn:
         conn.row_factory = model_factory(EmployeeComputer)
