+++ hrapp/views/employees/data_manager/computers/get_availible_computers.py
@@ -1,5 +1,5 @@
 import sqlite3
-from ...connection import Connection
+from hrapp.views.connection import Connection
 from hrapp.models import model_factory, EmployeeComputer, Computer
