+++ hrapp/views/employees/data_manager/departments/__init__.py
@@ -1,0 +1,1 @@
+from .get_all_departments import get_all_departments
