+++ hrapp/views/employees/data_manager/departments/get_all_departments.py
@@ -1,20 +1,20 @@
 import sqlite3
-from ...connection import Connection
-from hrapp.models import model_factory, TrainingProgram
+from hrapp.views.connection import Connection
+from hrapp.models import model_factory, Department
 
-def get_all_trainings():
+def get_all_departments():
     """
     This function gets all of the department data from hrapp_department
     """
     with sqlite3.connect(Connection.db_path) as conn:
-        conn.row_factory = model_factory(TrainingProgram)
+        conn.row_factory = model_factory(Department)
         db_cursor = conn.cursor()
 
         db_cursor.execute("""
             SELECT
                 *
             FROM
-                hrapp_trainingprogram
+                hrapp_department
         """)
 
         return db_cursor.fetchall()
