+++ hrapp/views/employees/data_manager/employees/get_employee.py
@@ -1,5 +1,5 @@
 import sqlite3
-from ...connection import Connection
+from hrapp.views.connection import Connection
 from hrapp.models import model_factory, Employee
 def get_employee(employee_id):
     """
