+++ hrapp/views/employees/data_manager/programs/__init__.py
@@ -1,0 +1,3 @@
+from .employee_trainings import get_employee_training
+from .all_trainings import get_all_trainings
+from .assign_program import assign_program
